clean up NcAudio
